Simplify style guide navigation
The current http://v4.style.codeforamerica.org has two levels of
navigation: a (empty) home page, and then when clicking into "Patten
Library" on the left side, you end up at the style guide itself.

This is also inconsistent branding. Let's remove the "pattern library"
terminology and just go with "Style Guide". This commit flattens out the
site so that there is only one layout -- and the navigation on the
sidebar is useful from the homepage.