Move `main.scss` into `_sass` directory
This will allow styleguide users to easily include the whole thing in
one go, like

@include '@codeforamerica/style/_sass/main';

(previously, the YAML frontmatter broke Sass when including it like so.)

Also, this has a pleasant quality of making the styleguide website
actually a client of the styleguide itself! This means we could break
the _sass directory into a different repo, if that would buy us
anything.