Add .info-alert component styling
Even though .info-alert is mentioned in the style guide, it's not
defined in the stylesheets. I need a very basic alert component to
convey information about a page, for example, 'This page describes an
past project.'. This commit adds a basic .info-alert component that
extends .alert and removes the bottom border.

We need to revisit the .alert component styling more holistically, but
for now I just want to use the lightest touch to add something very
basic.