Set background-size and background-position for hero elements
I'm adding the ability to specify a background image for header hero elements on the Code for
America website. This commit set some default properties for handling background images in hero
elements. Specifically, it makes the background image cover the hero element, and it positions the
background image 'center center'.