Add dark blue slab style
The stylesheet currently includes a blue background slab. A slight variation on the blue style would
be useful -- for example, the Summit 2018 branding uses medium blue and dark blue blocks next to
eachother. This commit updates `.blue-bg-slab` to use `$color-blue` instead of `$dark-blue`, and adds
a `.dark-blue-bg-slab` variant that uses `$color-dark-blue` for its background-color.

This commit adds a dark blue slab style.