Remove anchor tag selector from card styles
There are many reasons to not assume that an entire card is a link: if
you don't want the whole card to be a link, if you want to have multiple
links inside the card, and just in general if you want to use a card for
a non-linking purpose.

By removing the <a> from these selectors we can make the cards more
versatile.