Add a jumbo variant of the hero component
I'm adding a 'jumbo' variant of the hero component to be used in a new 'major events' template, for events like Summit. It has a bit of drama: it takes up most most (80%) of the vertical space of the screen and its content is displayed in its center.

This also add some styling of the sub-components of `hero--jumbo`, and starts including some BEM class name variants for the existing components.