Add column class names that start with 'col' followed by integer
In a 12 column grid, it's now pretty standard for the column classes to follow a naming convention that looks something like `col-3`, `col-6` etc. I want to introduce a bit more flexibility with the placement of components into grids in the CMS, and having the column class names in this more-standard format will make it easier to do so.