Add banner component
GivingTuesday is approaching and we want to add a prominent banner to the top of certain pages on the codeforamerica.org that includes an image that takes up half of the screen, along with a headline, some text, a call-to-action link and optionally some overhead text. This is a first stab at adding such a component.