Make the grid sizing of the grid row sub-element less agressive
Previously all elements in a grid-row column would size up to the full height of the column, but there are some situations where I want to stack two sub-elements in one column element and the alignment was getting a bit screwy. This fixes those alignment issues, and makes cards take up the full height of the row, so I can still do rows of cards that all the same height.