Add a subtle blockquote symbol to the left of blockquotes
There was some feedback that each blockquote should include a stylized quotation mark to make it clear that it's a quote. This does it by adding a background image of a transparent quotation mark.

There are ways to do it using only text, but the font we're using on the website currently has quotation marks that don't look good.

TODO: Add an 'inverted' version of the blockquote styling.