Use full URL for open-quotes image
The open quotes image isn't rendering in production. This adds the full URL for the image file, to make sure the image renders. This isn't an ideal implementation, but rather a quick fix. Ideally I'd like the quotes to be done using text only, and once we've updated the website fonts to Proxima Nova (for consistency with our other marketing materials), I'll change this out.