Darken gray shade to increase contrast between slabs
A common pattern on codeforamerica.org is to alternate white-bg and gray-bg slabs, but the contrast between the two backgrounds is quite low and one some devices (esp. on mobile) one might not even notice the variation in background colors. This darkens the gray-shade variable, used in gray-background-slabs, to the same shade that is used int global footer.