Remove recently-added blockquote styling
Follow some further feedback from the marketing team, this remove the styling added to blockquotes recently that displayed a quotation mark graphic to the left of blockquotes. The feeling is that only quotes attributable to people should include the mark, and should include a citation. Given that we're using the blockquote element for direct quotes *and* pull-quotes from articles, and given that the implementation wasn't best practice, I'm undoing those changes for now and wil revisit in the near future.