Fix code previews in style-guide
Code previews aren't showing up since I removed a class name from the style guide template in an earlier commit that was being used as a selector in the styleguide's JavaScript. This adds a different selector so that it works again.

The JavaScript could use a little rewriting -- I will return to it later.