Rename 'tag' component to 'label'
The 'tag' component is essentially a little label, and both Zurb Foundation and Semantic UI simple call it a 'label'. At the moment the style guide website uses the 'tag' class to style HTML tags, and the mixture of style makes them look very strange. Let's rename the 'tags' to 'labels' for clarity.