Update the way blockquote citations are handled
According to a few sources of online guidance (e.g. https://developer.mozilla.org/en-US/docs/Web/HTML/Element/cite) the right approach with blockquote references is to place them in a
<footer> element and to wrap the 'creative work' in a <cite> element.

As noted on the MDN web docs page: "...the W3C specification says that a reference to a creative work, as included within a <cite> element, may include the name of the work’s author. However, the WHATWG specification for <cite> says the opposite: that a person’s name must never be included, under any circumstances."

Let's stick to a convention where only the creative work itself is wrapped in <cite>.